+++ lib/presentation/home/component/recommend_ootd_widget.dart
@@ -2,8 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:weaco/domain/feed/model/feed.dart';
 import 'package:weaco/presentation/common/component/cached_image_widget.dart';
 
-class RecommandOotdWidget extends StatelessWidget {
-  const RecommandOotdWidget({
+class RecommendOotdWidget extends StatelessWidget {
+  const RecommendOotdWidget({
     super.key,
     required this.feedList,
     required this.index,
