+++ lib/presentation/home/screen/home_screen.dart
@@ -5,7 +5,7 @@ import 'package:weaco/presentation/common/style/image_path.dart';
 import 'package:weaco/core/enum/weather_code.dart';
 import 'package:weaco/presentation/common/enum/exception_alert.dart';
 import 'package:weaco/presentation/common/util/alert_util.dart';
-import 'package:weaco/presentation/home/component/recommand_ootd_list_widget.dart';
+import 'package:weaco/presentation/home/component/recommend_ootd_list_widget.dart';
 import 'package:weaco/presentation/home/component/weather_by_time_list_widget.dart';
 import 'package:weaco/presentation/home/view_model/home_screen_view_model.dart';
 
@@ -228,7 +228,7 @@ class _HomeScreenState extends State<HomeScreen> {
                     ),
 
                     // ootd list
-                    RecommandOotdListWidget(
+                    RecommendOotdListWidget(
                       dailyLocationWeather: viewModel.dailyLocationWeather,
                       feedList: viewModel.feedList,
                     ),
